Use double precision for rotation and precession periods. Orbital periods are doubles already, so switching rotation periods to doubles means that synchronous rotation will never drift due to precision differences.

+++ src/celengine/parseobject.h
@@ -30,7 +30,7 @@ Orbit* CreateOrbit(PlanetarySystem* system,
 
 RotationModel* CreateRotationModel(Hash* rotationData,
                                    const string& path,
-                                   float syncRotationPeriod);
+                                   double syncRotationPeriod);
 
 RotationModel* CreateDefaultRotationModel(double syncRotationPeriod);
